+++ CHANGELOG.md
@@ -5,5 +5,10 @@
 - Change package name `corrP` to `corrp` ;
 - Changelog file created ;
 - Fixed some bugs in function's documentations;
-- Add verbose param to corrp function ; 
-- 
+- Added verbose param to corrp function ; 
+- Added new correlations types analysis: pps ; dcor ; mic ; uncoef;
+- corrp function output has a new class `clist` with index matrix and data values.
+- corr_matrix: new function to create correlation matrix ;
+- corr_rm: Remove highly correlated variables from a data.frame 
+- Added testthat;
+- Passed in all checks;